requirements.txt
# Root requirements file
Scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
